+++ test/simulate/simulateDruid.mocha.js
@@ -13,6 +13,7 @@ var attributes = [
   { name: 'time', type: 'TIME' },
   { name: 'color', type: 'STRING' },
   { name: 'cut', type: 'STRING' },
+  { name: 'isNice', type: 'BOOLEAN' },
   { name: 'tags', type: 'SET/STRING' },
   { name: 'carat', type: 'NUMBER' },
   { name: 'height_bucket', special: 'range', separator: ';', rangeSize: 0.05, digitsAfterDecimal: 2 },
@@ -508,14 +509,14 @@ describe("simulate Druid", () => {
     });
   });
 
-  it("works with basic timePart", () => {
+  it("works with timePart (with limit)", () => {
     var ex = ply()
       .apply(
         'HoursOfDay',
         $("diamonds").split("$time.timePart(HOUR_OF_DAY, 'Etc/UTC')", 'HourOfDay')
           .apply('TotalPrice', '$diamonds.sum($price)')
-          .sort('$TotalPrice', 'descending')
-          .limit(3)
+          .sort('$HourOfDay', 'ascending')
+          .limit(20)
       );
 
     expect(ex.simulateQueryPlan(context)).to.deep.equal([
@@ -543,9 +544,95 @@ describe("simulate Druid", () => {
         "intervals": [
           "2015-03-12/2015-03-19"
         ],
-        "metric": "TotalPrice",
+        "metric": {
+          "type": "alphaNumeric"
+        },
         "queryType": "topN",
-        "threshold": 3
+        "threshold": 20
+      }
+    ]);
+  });
+
+  it("works with timePart (no limit)", () => {
+    var ex = ply()
+      .apply(
+        'HoursOfDay',
+        $("diamonds").split("$time.timePart(HOUR_OF_DAY, 'Etc/UTC')", 'HourOfDay')
+          .sort('$HourOfDay', 'ascending')
+      )
+      .apply(
+        'SecondOfDay',
+        $("diamonds").split("$time.timePart(SECOND_OF_DAY, 'Etc/UTC')", 'HourOfDay')
+          .sort('$HourOfDay', 'ascending')
+      );
+
+    expect(ex.simulateQueryPlan(context)).to.deep.equal([
+      {
+        "aggregations": [
+          {
+            "name": "!DUMMY",
+            "type": "count"
+          }
+        ],
+        "dataSource": "diamonds",
+        "dimension": {
+          "dimension": "__time",
+          "extractionFn": {
+            "format": "H",
+            "locale": "en-US",
+            "timeZone": "Etc/UTC",
+            "type": "timeFormat"
+          },
+          "outputName": "HourOfDay",
+          "type": "extraction"
+        },
+        "granularity": "all",
+        "intervals": [
+          "2015-03-12/2015-03-19"
+        ],
+        "metric": {
+          "type": "alphaNumeric"
+        },
+        "queryType": "topN",
+        "threshold": 1000
+      },
+      {
+        "aggregations": [
+          {
+            "name": "!DUMMY",
+            "type": "count"
+          }
+        ],
+        "dataSource": "diamonds",
+        "dimensions": [
+          {
+            "dimension": "__time",
+            "extractionFn": {
+              "format": "H'*60+'m'*60+'s",
+              "locale": "en-US",
+              "timeZone": "Etc/UTC",
+              "type": "timeFormat"
+            },
+            "outputName": "HourOfDay",
+            "type": "extraction"
+          }
+        ],
+        "granularity": "all",
+        "intervals": [
+          "2015-03-12/2015-03-19"
+        ],
+        "limitSpec": {
+          "columns": [
+            {
+              "dimension": "HourOfDay",
+              "dimensionOrder": "alphaNumeric",
+              "direction": "ascending"
+            }
+          ],
+          "limit": 500000,
+          "type": "default"
+        },
+        "queryType": "groupBy"
       }
     ]);
   });
@@ -844,6 +931,49 @@ describe("simulate Druid", () => {
         "metric": "Count",
         "queryType": "topN",
         "threshold": 10
+      }
+    ]);
+  });
+
+  it("works with BOOLEAN split", () => {
+    var ex = $("diamonds").split("$isNice", 'IsNice')
+      .apply('Count', $('diamonds').count())
+      .sort('$Count', 'descending');
+
+    expect(ex.simulateQueryPlan(context)).to.deep.equal([
+      {
+        "aggregations": [
+          {
+            "name": "Count",
+            "type": "count"
+          }
+        ],
+        "dataSource": "diamonds",
+        "dimension": {
+          "dimension": "isNice",
+          "extractionFn": {
+            "injective": false,
+            "lookup": {
+              "map": {
+                "0": "false",
+                "1": "true",
+                "false": "false",
+                "true": "true"
+              },
+              "type": "map"
+            },
+            "type": "lookup"
+          },
+          "outputName": "IsNice",
+          "type": "extraction"
+        },
+        "granularity": "all",
+        "intervals": [
+          "2015-03-12/2015-03-19"
+        ],
+        "metric": "Count",
+        "queryType": "topN",
+        "threshold": 1000
       }
     ]);
   });