+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change log
 
+## 0.8.14
+
+- DruidExternal will use topNs for defined small splits even when there is no limit set (like `BOOLEAN`, `timeBucket(HOUR_OF_DAY)`)
+- Fixed bug for MySQLExternal not dealing well with BOOLEAN splits 
+
 ## 0.8.13
 
 - Added `overlap` action